Backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Werkzeug==3.0.2
langchain==0.1.20
langchain-community==0.0.38
faiss-cpu==1.8.0.post1
pdfminer.six==20231228
openai==1.51.0
PyMuPDF==1.24.1
gunicorn
tiktoken
python-dotenv